program class reestructured
